Module malachite_base::chars::constants

Constants associated with chars.

Apart from the constants visibile on this page, the trait-based constants MIN, MAX, and NAME are also defined.

Constants

CHAR_JUST_ABOVE_SURROGATES

The char that comes just after the surrogate range.

CHAR_JUST_BELOW_SURROGATES

The char that comes just before the surrogate range.

FIRST_SURROGATE_CODE_POINT

The first surrogate code point.

NUMBER_OF_CHARS

The number of Unicode scalar values.

NUMBER_OF_SURROGATE_CODE_POINTS

The number of surrogate code points; these are code points that do not correspond to any valid char.