Module malachite_base::iterators::comparison

Functions that compare adjacent iterator elements.

Structs

DeltaDirections

An iterator that generates the Orderings of adjacent elements of a given iterator.

Functions

delta_directions

Returns an iterator that generates the Orderings of adjacent pairs of elements of a given iterator.

is_strictly_ascending

Determines whether each element of an iterator is greater than the preceding one.

is_strictly_descending

Determines whether each element of an iterator is less than the preceding one.

is_strictly_zigzagging

Determines whether the sequence strictly zigzags.

is_weakly_ascending

Determines whether each element of an iterator is greater than or equal to the preceding one.

is_weakly_descending

Determines whether each element of an iterator is less than or equal to the preceding one.

is_weakly_zigzagging

Determines whether the sequence weakly zigzags.