Trait malachite_base::num::arithmetic::traits::CeilingDivAssignMod

pub trait CeilingDivAssignMod<RHS = Self> {
    type ModOutput;

    // Required method
    fn ceiling_div_assign_mod(&mut self, other: RHS) -> Self::ModOutput;
}

Divides a number by another number in place, taking the quotient and returning the remainder. The quotient is rounded towards positive infinity and the remainder has the opposite sign of the divisor (second input).

The quotient and remainder satisfy $x = qy + r$ and $0 \leq |r| < |y|$.

Required Associated Types

type ModOutput

Required Methods

fn ceiling_div_assign_mod(&mut self, other: RHS) -> Self::ModOutput