Trait malachite_base::num::arithmetic::traits::RoundToMultipleAssign

pub trait RoundToMultipleAssign<RHS = Self> {
    // Required method
    fn round_to_multiple_assign(
        &mut self,
        other: RHS,
        rm: RoundingMode
    ) -> Ordering;
}

Rounds a number to a multiple of another number in place, according to a specified rounding mode. Ordering is returned, indicating whether the returned value is less than, equal to, or greater than the original value.

Required Methods

fn round_to_multiple_assign(&mut self, other: RHS, rm: RoundingMode) -> Ordering