Trait malachite_base::num::arithmetic::traits::ShrRound

pub trait ShrRound<RHS> {
    type Output;

    // Required method
    fn shr_round(self, other: RHS, rm: RoundingMode) -> (Self::Output, Ordering);
}

Right-shifts a number (divides it by a power of 2), rounding the result according to a specified rounding mode. An Ordering is also returned, indicating whether the returned value is less than, equal to, or greater than the exact value.

Rounding might only be necessary if other is positive.

Required Associated Types

type Output

Required Methods

fn shr_round(self, other: RHS, rm: RoundingMode) -> (Self::Output, Ordering)