Trait malachite_base::num::arithmetic::traits::SqrtRem

pub trait SqrtRem {
    type SqrtOutput;
    type RemOutput;

    // Required method
    fn sqrt_rem(self) -> (Self::SqrtOutput, Self::RemOutput);
}

Finds the floor of the square root of a number, returning both the root and the remainder.

Required Associated Types

type SqrtOutput

type RemOutput

Required Methods

fn sqrt_rem(self) -> (Self::SqrtOutput, Self::RemOutput)