Trait malachite_base::num::arithmetic::traits::XXXSubYYYToZZZ

pub trait XXXSubYYYToZZZ: Sized {
    // Required method
    fn xxx_sub_yyy_to_zzz(
        x_2: Self,
        x_1: Self,
        x_0: Self,
        y_2: Self,
        y_1: Self,
        y_0: Self
    ) -> (Self, Self, Self);
}

Subtracts two numbers, each composed of three Self values, returing the difference as a triple of Self values.

The more significant number always comes first. Subtraction is wrapping, and overflow is not indicated.

Required Methods

fn xxx_sub_yyy_to_zzz( x_2: Self, x_1: Self, x_0: Self, y_2: Self, y_1: Self, y_0: Self ) -> (Self, Self, Self)

Object Safety

This trait is not object safe.