Trait malachite_base::num::comparison::traits::PartialOrdAbs

pub trait PartialOrdAbs<Rhs: ?Sized = Self> {
    // Required method
    fn partial_cmp_abs(&self, other: &Rhs) -> Option<Ordering>;

    // Provided methods
    fn lt_abs(&self, other: &Rhs) -> bool { ... }
    fn le_abs(&self, other: &Rhs) -> bool { ... }
    fn gt_abs(&self, other: &Rhs) -> bool { ... }
    fn ge_abs(&self, other: &Rhs) -> bool { ... }
}

Determines equality between the absolute values of two numbers, where some pairs of numbers may not be comparable.

Required Methods

fn partial_cmp_abs(&self, other: &Rhs) -> Option<Ordering>

Compares the absolute values of two numbers, taking both by reference.

If the two values are not comparable, None is returned.

Provided Methods

fn lt_abs(&self, other: &Rhs) -> bool

Determines whether the absolute value of one number is less than the absolute value of another.

Worst-case complexity

Same as the time and additional memory complexity of partial_cmp_abs.

fn le_abs(&self, other: &Rhs) -> bool

Determines whether the absolute value of one number is less than or equal to the absolute value of another.

Worst-case complexity

Same as the time and additional memory complexity of partial_cmp_abs.

fn gt_abs(&self, other: &Rhs) -> bool

Determines whether the absolute value of one number is greater than the absolute value of another.

Worst-case complexity

Same as the time and additional memory complexity of partial_cmp_abs.

fn ge_abs(&self, other: &Rhs) -> bool

Determines whether the absolute value of one number is greater than or equal to the absolute value of another.

Worst-case complexity

Same as the time and additional memory complexity of partial_cmp_abs.