Struct malachite_base::num::conversion::string::to_string::BaseFmtWrapper

pub struct BaseFmtWrapper<T> { /* private fields */ }

A struct that allows for formatting a numeric type and rendering its digits in a specified base.

Implementations

impl<T> BaseFmtWrapper<T>

pub fn new(x: T, base: u8) -> Self

Creates a new BaseFmtWrapper.

Worst-case complexity

Constant time and additional memory.

Panics

Panics if base is less than 2 or greater than 36.

Examples

use malachite_base::num::conversion::string::to_string::BaseFmtWrapper;

let x = BaseFmtWrapper::new(1000000000u32, 36);
assert_eq!(format!("{}", x), "gjdgxs");
assert_eq!(format!("{:#}", x), "GJDGXS");

pub fn unwrap(self) -> T

Recovers the value from a BaseFmtWrapper.

Worst-case complexity

Constant time and additional memory.

Examples

use malachite_base::num::conversion::string::to_string::BaseFmtWrapper;

assert_eq!(BaseFmtWrapper::new(1000000000u32, 36).unwrap(), 1000000000);