Trait malachite_base::num::conversion::traits::ConvertibleFrom

pub trait ConvertibleFrom<T> {
    // Required method
    fn convertible_from(value: T) -> bool;
}

Tests whether a value of one type is convertible into a value of another.

If ConvertibleFrom<T> for Self is implemented, it usually makes sense to implement TryFrom for T as well.

Required Methods

fn convertible_from(value: T) -> bool

Object Safety

This trait is not object safe.