Trait malachite_base::num::conversion::traits::OverflowingFrom

pub trait OverflowingFrom<T>: Sized {
    // Required method
    fn overflowing_from(value: T) -> (Self, bool);
}

Converts a value from one type to another, where if the conversion is not exact the result will wrap around. The result is returned along with a bool that indicates whether wrapping has occurred.

If OverflowingFrom is implemented, it usually makes sense to implement WrappingFrom as well.

Required Methods

fn overflowing_from(value: T) -> (Self, bool)

Object Safety

This trait is not object safe.