Trait malachite_base::num::conversion::traits::PowerOf2Digits

pub trait PowerOf2Digits<T>: Sized {
    // Required methods
    fn to_power_of_2_digits_asc(&self, log_base: u64) -> Vec<T>;
    fn to_power_of_2_digits_desc(&self, log_base: u64) -> Vec<T>;
    fn from_power_of_2_digits_asc<I: Iterator<Item = T>>(
        log_base: u64,
        digits: I
    ) -> Option<Self>;
    fn from_power_of_2_digits_desc<I: Iterator<Item = T>>(
        log_base: u64,
        digits: I
    ) -> Option<Self>;
}

Expresses a value as a Vec of base-$2^k$ digits, or reads a value from an iterator of base-$2^k$ digits.

The trait is parameterized by the digit type.

Required Methods

fn to_power_of_2_digits_asc(&self, log_base: u64) -> Vec<T>

Returns a Vec containing the digits of a value in ascending order: least- to most-significant.

The base is $2^k$, where $k$ is log_base.

fn to_power_of_2_digits_desc(&self, log_base: u64) -> Vec<T>

Returns a Vec containing the digits of a value in descending order: most- to least-significant.

The base is $2^k$, where $k$ is log_base.

fn from_power_of_2_digits_asc<I: Iterator<Item = T>>( log_base: u64, digits: I ) -> Option<Self>

Converts an iterator of digits into a value.

The input digits are in ascending order: least- to most-significant. The base is $2^k$, where $k$ is log_base.

fn from_power_of_2_digits_desc<I: Iterator<Item = T>>( log_base: u64, digits: I ) -> Option<Self>

Converts an iterator of digits into a value.

The input digits are in descending order: most- to least-significant. The base is $2^k$, where $k$ is log_base.

Object Safety

This trait is not object safe.