Trait malachite_base::num::conversion::traits::RawMantissaAndExponent

pub trait RawMantissaAndExponent<M, E, T = Self>: Sized {
    // Required methods
    fn raw_mantissa_and_exponent(self) -> (M, E);
    fn from_raw_mantissa_and_exponent(raw_mantissa: M, raw_exponent: E) -> T;

    // Provided methods
    fn raw_mantissa(self) -> M { ... }
    fn raw_exponent(self) -> E { ... }
}

Converts a number to or from a raw mantissa and exponent.

See here for a definition of raw mantissa and exponent.

Required Methods

fn raw_mantissa_and_exponent(self) -> (M, E)

Extracts the raw mantissa and exponent from a number.

fn from_raw_mantissa_and_exponent(raw_mantissa: M, raw_exponent: E) -> T

Constructs a number from its raw mantissa and exponent.

Provided Methods

fn raw_mantissa(self) -> M

Extracts the raw mantissa from a number.

fn raw_exponent(self) -> E

Extracts the raw exponent from a number.

Object Safety

This trait is not object safe.