Struct malachite_base::num::float::NiceFloat

pub struct NiceFloat<T: PrimitiveFloat>(pub T);

NiceFloat is a wrapper around primitive float types that provides nicer Eq, Ord, Hash, Display, and FromStr instances.

In most languages, floats behave weirdly due to the IEEE 754 standard. The NiceFloat type ignores the standard in favor of more intuitive behavior.

Using NiceFloat, NaNs are equal to themselves. There is a single, unique NaN; there’s no concept of signalling NaNs. Positive and negative zero are two distinct values, not equal to each other.

The NiceFloat hash respects this equality.

NiceFloat has a total order. These are the classes of floats, in ascending order:

Negative infinity

Negative nonzero finite floats

Negative zero

NaN

Positive zero

Positive nonzero finite floats

Positive infinity

NiceFloat uses a different Display implementation than floats do by default in Rust. For example, Rust will format f32::MIN_POSITIVE_SUBNORMAL as something with many zeros, but NiceFloat(f32::MIN_POSITIVE_SUBNORMAL) just formats it as "1.0e-45". The conversion function uses David Tolnay’s ryu crate, with a few modifications:

All finite floats have a decimal point. For example, Ryu by itself would convert f32::MIN_POSITIVE_SUBNORMAL to "1e-45".

Positive infinity, negative infinity, and NaN are converted to the strings "Infinity", "-Infinity", and “NaN”, respectively.

FromStr accepts these strings.

Tuple Fields
0: T