Trait malachite_base::num::logic::traits::BitAccess

pub trait BitAccess {
    // Required methods
    fn get_bit(&self, index: u64) -> bool;
    fn set_bit(&mut self, index: u64);
    fn clear_bit(&mut self, index: u64);

    // Provided methods
    fn assign_bit(&mut self, index: u64, bit: bool) { ... }
    fn flip_bit(&mut self, index: u64) { ... }
}

Defines functions that access or modify individual bits in a number.

Required Methods

fn get_bit(&self, index: u64) -> bool

Determines whether a number has a true or false bit at index.

fn set_bit(&mut self, index: u64)

Sets the bit at index to true.

fn clear_bit(&mut self, index: u64)

Sets the bit at index to false.

Provided Methods

fn assign_bit(&mut self, index: u64, bit: bool)

Sets the bit at index to whichever value bit is.

Worst-case complexity

$T(n) = O(\max(T_S(n), T_C(n)))$,

$M(n) = O(\max(M_S(n), M_C(n)))$

where $T$ is time, $M$ is additional memory, $T_S$ and $M_S$ are the complexities of set_bit, and $T_C$ and $M_C$ are the complexities of clear_bit.

Panics

See panics for set_bit and clear_bit.

fn flip_bit(&mut self, index: u64)

Sets the bit at index to the opposite of its original value.

Worst-case complexity

$T(n) = O(T_G(n) + \max(T_S(n), T_C(n)))$,

$M(n) = O(M_G(n) + \max(M_S(n), M_C(n)))$

where $T$ is time, $M$ is additional memory, $T_G$ and $M_G$ are the complexities of get_bit, $T_S$ and $M_S$ are the complexities of set_bit, and $T_C$ and $M_C$ are the complexities of clear_bit.

Panics

See panics for get_bit, set_bit and clear_bit.