Trait malachite_base::num::logic::traits::BitConvertible

pub trait BitConvertible {
    // Required methods
    fn to_bits_asc(&self) -> Vec<bool>;
    fn to_bits_desc(&self) -> Vec<bool>;
    fn from_bits_asc<I: Iterator<Item = bool>>(bits: I) -> Self;
    fn from_bits_desc<I: Iterator<Item = bool>>(bits: I) -> Self;
}

Defines functions that express a number as a Vec of bits or construct a number from an iterator of bits.

Required Methods

fn to_bits_asc(&self) -> Vec<bool>

Returns a Vec containing the bits of a number in ascending order: least- to most-significant.

fn to_bits_desc(&self) -> Vec<bool>

Returns a Vec containing the bits of a number in descending order: most- to least-significant.

fn from_bits_asc<I: Iterator<Item = bool>>(bits: I) -> Self

Converts an iterator of bits into a number. The input bits are in ascending order: least- to most-significant.

fn from_bits_desc<I: Iterator<Item = bool>>(bits: I) -> Self

Converts an iterator of bits into a value. The input bits are in descending order: most- to least-significant.

Object Safety

This trait is not object safe.