Function malachite_base::options::exhaustive::exhaustive_somes

pub const fn exhaustive_somes<I: Iterator>(xs: I) -> ExhaustiveSomes<I> 

Generates all Options except None, with values from a given iterator.

The elements of the given iterator are wrapped in Some and generated in the original order.

The output length is xs.count().

Examples

use itertools::Itertools;
use malachite_base::options::exhaustive::exhaustive_somes;

assert_eq!(
    exhaustive_somes([1, 2, 3].iter().cloned()).collect_vec(),
    &[Some(1), Some(2), Some(3)]
);