Module malachite_base::strings

Functions for working with Strings.

Modules

exhaustive

Iterators that generate Strings without repetition.

Structs

StringsFromCharVecs

Generates Strings, given an iterator that generates Vec<char>s.

Traits

ToBinaryString

A trait that provides an ergonomic way to create the string specified by a Binary implementation.

ToDebugString

A trait that provides an ergonomic way to create the string specified by a Debug implementation.

ToLowerHexString

A trait that provides an ergonomic way to create the string specified by a LowerHex implementation.

ToOctalString

A trait that provides an ergonomic way to create the string specified by an Octal implementation.

ToUpperHexString

A trait that provides an ergonomic way to create the string specified by an UpperHex implementation.

Functions

string_is_subset

Returns whether all of the first string slice’s characters are present in the second string slice.

string_sort

Sorts the characters of a string slice and returns them in a new String.

string_unique

Takes a string slice’s unique characters and returns them in a new String.

strings_from_char_vecs

Generates Strings, given an iterator that generates Vec<char>s.