Trait malachite_base::num::conversion::traits::SciMantissaAndExponent

pub trait SciMantissaAndExponent<M, E, T = Self>: Sized {
    // Required methods
    fn sci_mantissa_and_exponent(self) -> (M, E);
    fn from_sci_mantissa_and_exponent(
        sci_mantissa: M,
        sci_exponent: E
    ) -> Option<T>;

    // Provided methods
    fn sci_mantissa(self) -> M { ... }
    fn sci_exponent(self) -> E { ... }
}

Converts a number to or from a scientific mantissa and exponent.

See here for a definition of scientific mantissa and exponent.

Required Methods

fn sci_mantissa_and_exponent(self) -> (M, E)

Extracts the scientific mantissa and exponent from a number.

fn from_sci_mantissa_and_exponent(sci_mantissa: M, sci_exponent: E) -> Option<T>

Constructs a number from its scientific mantissa and exponent.

Provided Methods

fn sci_mantissa(self) -> M

Extracts the scientific mantissa from a number.

fn sci_exponent(self) -> E

Extracts the scientific exponent from a number.

Object Safety

This trait is not object safe.